Create my-last-index-of.js and README.md
Function accepts an array, a search value, and an optional starting index value variable.  Uses search value to search the array and return the location of the value in the array.  If starting index is specified, searches from that location in the array to the end.  If the value does not exist in the array, returns -1.